Divide denominators by a common divisor before operating on fractions.
If numbers p1/q1 and p2/q2 are added, then it makes sense to remove
the greatest common divisor of q1 and q2 before adding the numbers.
Suppose q1 and q2 have a gcd n, i.e. q1=n*r1 and q2=n*r2, then
adding p1/q1 and p2/q2 is

       p1*r2 + p2*r1
       -------------
         r1*r2*n

instead of the more complex

       p1*q2 + p2*q1
       -------------
           q1*q2

E.g adding 17/256  and 15/128 has a gcd of 128. The result is

       17*1 + 15*2
       -----------
          2*1*128

which is easier to calculate than

       17*128 + 15*256
      -----------------
         256*128