Fixed an omission in ltspmerge indicated by new tests.
The probabilistic states have been altered a while ago
to contain either a state, or contain a vector with
2 or more states with associated probability. Originally,
this was always a vector, but generating a vector for
every probabilistic state with only one element was
very time consuming. Therefore, this had been altered.

Unfortunately, this change was not properly applied
in the code for merging two probabilistic transition systems.
This has now been mended.