Removed an error in rewriting const_multiply.
In const_multiply 0*phi should be 0.
However in expressions of the shape if(cond,1,0)*phi
it was erroneously not recognized that the first expression
could be zero. In case phi was infinite, i.e., true, the expression
would simplify to true, which is only correct if the condition is
valid.

Added a few tests to the pres framework. More tests are required.