Repaired reading large size_t's from an aterm stream.
A large integer on an aterm is stored in chunks of 7 bits,
and could be stored in 10 bytes at times.
When reading such a number, reading would throw an exception
if the representation would exceed the length of 8 bytes.
This meant that in some cases large numbers would not be properly
read.