Resolves a race condition when using caches.
As indicated in bug report #1763 there is a problem
when between a find of an unordered map, and the comparison
of the result of this find, the end() of the unordered_map has
changed. By puttin explicit locks and unlocks around it the
problem is solved. With stress testing, with an explicit
sleep between the find and the end, the problem does not
occur anymore.

This closes #1763.