Aligned two inconsistent iteration limits the data rewriter.
When universal and existential quantifiers are enumerated in
data rewriters over unbounded data domains the number of elements
in that domain is limited. I.e., by default forall n:Nat.phi
cannot be expanded for more than 10 values of n. Unfortunately,
there were two mechanisms to guarantee this. One in the data rewriter,
which fixed this number to 10, and one in the iterating enumerator
which used a value, by default 1000, that could be set with the -Q
or --qlimit flag in rewriter tools.

The explicit rewriter limit has now been removed, and the
built in default value of the iterating enumerator has now been
set to 10.

This should not affect default behaviour. The only noticable behaviour
is that the flag --qlimit works again.

All tests succeed.