Reenabled the obligatory val in modalities.
By mistake I believed that val expressions in modalities
in modal formulas were never required, but that they accidentally
were inserted when introducing the pres framework.

However, this was a mistake. Now the val has to be used
also around data expressions inside modal formulas.

A long standing problem is still to remove the val everywhere.

In this commit, the use of val inside modalities has not
been adapted, and therefore tests may fail.