Resolved a failing examples test.
The sultan of persia examples test failed in debug mode.
The problem turned out to be that the initial pres variable was
not normalized. This meant that substitutions not in normal form
were entered. As the rewriter asssumes that terms are in normal
form, this caused the rewriter to not always work properly.

Also removed som superfluous code from the pres simplify_rewriter.