Resolve bug report #1778.
This bug was due to applying lps2pbes to stochastic LPS's.
This is not defined, but it was not noticed. This has now been resolved
by explicitly throwing an exception.

There were two sets of functions called load_lps, and save_lps. One
of them has been removed.

The message "Loading LPS" and "Saving LPS" is now only printed in
debug mode, to avoid excessive messages on the screen.

One tooltest fails, but appears reasonable, and still has
to be resolved.