Finished a first version of the experimental GJ branching bisimulation algorithm.
Under a compilation flag, added the new algorithm to list of
equivalences, and made it possible to start experimenting.
Without setting the flag, the behaviour of the tools are not altered.

The experimental algorithm does not compile, but this is not noticable
if the flag #define BRANCH_BIS_EXPERIMENT_JFG is not set.