Use compiled tools for example tests
The example tests run with the tools first found in the configured PATH.
Currently, if e.g. a system-wide mCRL2 install is present, the tests are
run against this installation instead of the compiled version.

Prepend the tool path to PATH instead of appending it to make sure that
the compiled tools are found before the system-wide install.