Added an apparently working version of a new branching bisimulation algorithm
The new implementation for branching bisimulation now appears to work.
Added it to the conversion tools, where it can be used using the flag
branching-bisim-gj. It is in no sense a fast implementation, and it
outputs lots of debug information when applied.

Also added/revamped a few tests for reductions. All tests succeed.