Substantially improved version of the branching bisimulation algorithm.
The performance is increased a 100 fold, by properly implementing the
co-routines such that they do a constant amount of work while splitting
the marked and unmarked states. The performance now typically is in
the order of a few times (ie. four times in a large case) slower than
the jgkw algorithm.