Further refined the new branching bisimulation algorithm.
As it stands the algorithm passes many random tests.
However, the core algorithm is typically 3 times slower than
jgkw. Based on a number of benchmarks, it appears that this
is primarily due to the implementation style, which can be
optimised rather than the algorithm itself.

For instance on a large variant of the 1394 protocol with 20 data
elements (43M states, 39 after SCC reduction, and 100M transitions)
we get:

                   gj (new algorithm)  jgkw
SCC                45s                   45s
initialisation     36s                    5s
refinement         54s                   33s
finalizing         27s                   24s

In the gj algorithm initialisation also partitions the states based on
its outgoing signature. However, more than 50% of the time is spent
in setting up the data structures, which takes linear time (appr. 20s).
Finalizing implies the generation of a reduced  transition system
based on the generated partition.