Removed the unordered_map to repair the start_same_saC pointers.
Recalling which start_same_saC pointers had to be repaired was
done with an unordered map, mapping pairs to another pair. This
could be replaced by a single, simple vector. The result is
a speed up of 10% of the total running time.

Extensive tests succeed, showing that the algorithm is correct.