Constructing very long lists did not work out well.
Using make_term_list for lists longer than 10.000 elements, the
list was not terminated with an empty list. Terms containing large
lists, as occurring in the example in bug report #1757 got corrupted.
Apparently, not too many of such large lists are being constructed,
given that this problem only appears to emerge now.