Adapted the representative_generator to yield a deterministic result.
Using a representative generator it is possible to generate a term
of some given sort. Up till now this term was generated based on
how terms were stored in memory. This meant that in different runs
of the tools different terms could be generated. This was annoying,
as various runs of the simulator could have different terms, which
would mean that generated traces were not always exchangeable between
different runs.

This has now been mended. When there is a choice the lexicographically
minimal function symbol is chosen. Note that there is still a preference
for constants over more complex expressions, and a preference of
mapping constants over constructor constants.

All tests succeed.

This addresses issue #1743.