Extended pretty printing to untyped operators on numbers.
In bug report #1735 it is indicated that printing a numerical
expression such as (x-1)*y in an error message drops the brackets,
leading to the confusing x-1*y. The reason is that the pretty print
framework did not take precedences of untyped operators into
account. As typing failed -- leading to the error message -- the
operators - and * were still untyped.

This has now been resolved by extending the precedence framework
to untyped symbols *, -, + and /. It may turn out to be desired
to extend the framework to more symbols later, but most symbols,
such as ++, |> etc. only occur in contexts where they are already
fully typed when being parsed.

All tests succeed.