Adapted generating the transition system after reduction.
Took care that the generation of the transition system
uses the internal data structure, instead of walking through
all transitions. This has as an effect that the generation
is faster, but in no way as fast as I expected. Investigating
a factor 40 less transitions only leads to a speed up of a factor
3....