Two changes in liblts_bisim_gj.h.
Change 1 regards the optimalisation of swapping states in
m_states_in_blocks. This code is now cleaner.

Change 2 regards properly initialising count_transitions_per_label.
The space for this vector was reserved, but not assigned. On a Mac
this did not lead to problems, as apparently the memory is still
initialised, but Ubuntu could not handle this.