Some further optimisation of liblts_bisim_gj.h.
Swapping states in m_states_in_blocks has been cleaned up a little,
such that the case distinctions are now quite natural.

When marking states for a VARIANT 1 split the marked bottom states
are now put in front of the bottom states. The non-marked bottom states
are put at the end. This optimises walking through the non-marked
bottom states when calculating the set U of unmarked states.

For the 1394-fin-vvlarge example this leads to a marginal performance
improvement.