Time measurement for new bottom states
While adding time complexity measurement for calculations related to new bottom states,
it became apparent that some parts of these calculations take too much time;
I started to change W_empty() to accommodate for that, and I introduced a pointer
to the transitions starting in non-bottom states in a BLC-set.
However, I had to undo my changes during the merge with JFG's recent edits.

The problem with the current version of W_empty() is this: if a new bottom state
has a transition with the correct label to the splitter, the work done to move
this state to R can be assigned to the transition; however, if a new bottom state
has no such transition, one is not allowed to spend time on it. My original idea
was to only spend time on the R-states and handle the U-states later, during
the two coroutines; however, JFG seems to have undone my change.

More changes will be required, in particular a clear separation between old
and new bottom states. Currently it may still happen that a block with
new bottom states is stabilized using a splitter that contains no transitions
from new bottom states at all; the work done in such a step cannot be accounted for.