Solved #1786, where labels in a .aut file were renamed to multi-actions.
In bug report #1786 it was indicated that an action in a .aut file of
the shape a(arg) b would be transformed into a(arg)|b, changing an
action into a multi action. This has now been adapted by only generating
multi-actions when the symbol "|" is explicitly used in the input.

Added a test.

Repaired the lts_test, as failing tests were not reported.