Corrected bug in the newest optimization
The previous change of the GJ algorithm would check whether a BLC set was actually
split into two parts, because only in that case the two BLC sets need to be
handled as splitters. However, the check wrongly assumed that the formerly
inert transitions need not be handled as splitter. This version corrects the
latter error.

It also contains an additional test in ltsconvert_test that exposed the bug.