Added a hidden option add_state_as_state_label to ltsconvert.
Added the hidden flag --add_state_as_state_label to ltsconvert.
This option allows to add the current state as the state label
to ltsconvert in .lts and .fsm files. This can be useful to
track how states are grouped together when applying a branching
bisimulation reduction, as all state labels of all equivalent states
are then joined to form a list of state labels of the reduced state.