Minor adaptations in the new hidden flag of ltsconvert.
After experimenting with Mieke Massink, it was decided that the
flag --add_state_as_state_label should be renamed into --add-state-as-state-label
and moreover, when using the flag, it was forgotten to change the
parameter names of an .lts input file to state_number of sort Nat. This
has now been mended.