Cache process equations in alphabet reductions
Alphabet reductions performed repeated linear searches in the vector of equations to find a process identifier. In the worst case, every left-hand-side of an equation was searched for at least once.

We now, instead, precompute a map of process identifier to equations, so the cost of lookup becomes O(log(n)) instead of O(n) where n is the number of equations.

In my example with 413 equation, this leads to a 15% performance improvement.