Added a few hash functions to experiment with hashing aterm pointers.
The different hash functions work very different under different
circumstances. The "standard" hash function n>>4 is the fastest
when generating state spaces, but it is also sometimes dreadfully slow.
The others are generally slower, but hash much better.

As yet it is unclear what a good hash function looks like.