Bug corrected: handling of formerly inert transitions
A formerly constellation-inert transition was sometimes handled as if it
required to stay in a certain relationship with the corresponding co-splitter
(which still contains constellation-inert transitions). Searching for
this co-splitter led occasionally to errors.

This commit corrects the error (such formerly constellation-inert transitions
can be treated as transitions that are not in a special main splitter)
and adds a test case to ltsconvert_test.