Reinstated the old hash function for terms, and added a warning in the comments.
It turns out that the hash-function ">>4", i.e., shift four bits to the
right is very effective for aterms when used in chaining hash tables.
As all aterms are consecutive in memory, they more or less appear
consecutively in chaining hash tables. However, for those places where
hash tables with open addressing is used, this hash performs very
badly. After a number of experiments we have decided that we leave
the current situation as is, as replacing the ">>4" hash with a very
good "cityhash" or other hash function leads to a performance drop
of 30%, most likely due to worse caching behaviour.

However, in some cases, the ">>4" hash performs very badly, leading
the toolset to come to a standstill. So, at some future time, it may
turn out to be wise to switch to a better built in hash function.