Bug on stabilized splitters corrected
When a split under a main splitter is executed, the main splitter and the
corresponding co-splitter need to maintain their relationship, so that the
subsequent co-split will be executed correctly.  This relationship was
not always maintained correctly.  The current version improves on this
(and also tries to not maintain the relationship when it's not needed
any more, namely when a split under a co-splitter is executed).