Formatting and comments, first half
To see more clearly which parts of the code are essential for the algorithm
and which parts only serve for debugging, the latter parts are moved to
columns 80-160. If you are only interested in the essential algorithms,
you may concentrate on the first 80 characters of every line of code.

The above holds for (most of) the first half of the file now. I will
handle the rest later.