Debugged two versions of the experimental GJ 2024 branching bisimulation algorithm
This commit contains two (as it now appears) correct versions of initialisation:
depending on #define INITIAL_PARTITION_WITHOUT_BLC_SETS, the initial partition
is constructed as in the earlier version, or with a simplified refinement
algorithm that does not use BLC sets. The latter is hoped to be faster--
a similar simplified refinement algorithm was faster in a very old version
of the algorithm.