Iterator past end dereferences removed
Dereferencing the end() iterator of a vector `(even in "&*end()") is undefined
behaviour; to circumvent this problem, I first used std::next(&back()),
but that only works if the vector is not empty. I now use data()+size()
(wrapped in a function of fixed_vector named data_end()) to avoid these
problems. The addition is optimized away by my compiler.

Also cleaned up the code and formatted it in a few places.