Bug corrected + comments cleanup
There was a bug that sometimes a state would be marked as belonging to
MultiSub but it wouldn't be added to the vector of MultiSub states
(in a situation when MultiSub was already aborted, so it would not make
sense to add states to the vector).  This bug is now corrected:  in
some cases, it is not even necessary to mark the state as belonging
to MultiSub; in other cases, the state is marked and added.

Additionally removed a few debugging statements.