Dramatically improve the performance of communication.
In case there were many matches of communication expressions in a single summand, first all possible combinations of matches were explored, and the corresponding condition and summand generated. So, for a single summand that matches n communication expressions, 2^n possible multiactions are generated. In practice, however, many of these multiactions contain actions that are either blocked, or that are not part of any multiaction in an allow set.

I added two test cases that were very slow without this change, but that take hardly any time now. These cases are adapted from the mCRL2 translation of a Cordis model, for which without the modification mcrl22lps --no-alph takes 1550s, and with the modification it terminates in 11s.