Added an experimental tool with the name pbesbackelm.
This tool takes a pbes and substitutes each equation

sigma X(....)=phi

in the right hand sides of the equation for X and all earlier
equations. If phi simplifies by allowing to replace a propositional
variable instantiation by an expression without a variable, then
this replacement is carried out.

This may sometimes simplify equations, which make solving
easier.