Refactor statistics for communication
This is a preparatory step to more systematically print information about the operations in the linearizer, and the operations that are applied to it.

The plan is to log all the operations in a similar way, and generate a list of the operations and statistics in YAML format. We can see where intermediate results blow up and shrink immeidately afterwards to inspire optimizations that combine operations.