Complete revamp of the jittyc rewriter.
The jittyc rewriter did not always correctly substitute values
for variables bound in where clauses. In order to deal with this
the rewriter has been made more sequential, in the sense that
it evaluates subexpressions first, assign the result to the
rewrite stack, and subsequentially evaluate the main expression.

Tests have been added to show where rewriting of where clauses
went astray.

As this is a substantial overhaul of the compiling rewriter its
performance may have changed. It is expected to be much faster
with where clauses, and slightly faster overall. This is consistent
with the benchmarks.

All tests succeed.