Corrected bug with iterator checks
An iterator was (possibly) invalidated, then swapped, then set to the correct value;
this led to bugs when safe iterators are used. The order has been changed to
swapping, then setting to the correct value, then invalidating the former
value of the iterator (by deleting the object pointed at previously).