Removed an obvious odd piece of code in Martens' bisimulation algorithm.
In the code in liblts_bisim_minimal_depth.h an iteration was taking
place over a vector of indices BL, and this in the loop BL was cleared.
This meant that iteration went astray, being detected by safe iterators.
Most likely the clearing of BL was positioned at a wrong place in
the code. By outcommenting it, all strict test, using
MCRL2_ENABLE_STD_CHECKS succeed on the Mastodont.

Jan Martens should check what the intention of the code was, and
whether the clear could indeed be outcommented.