Correction of a bug reported by Maurice Laveaux involving singular iterators
If an iterator is swapped with itself using std::swap(), the move constructor
(for the temporary variable) will leave the original iterator in a singular
state; then, the move assignment operator will complain in debug mode that
a singular iterator is assigned to something.

This commit adds a check to ensure that an iterator is not swapped with itself.
Also added a few assert() expressions to check that I've understood the usage
correctly.