Repaired ltsgraph.
Ltsgraph had become quite unusable for unclear reasons.

Inspection of the code revealed that a fast_random function
was added that added huge random values to every pair of states
to repel or attract these state at will. From the code it was
clear that the intention was to only a small variation to
the repelling force, in the interval of [-0.1,0.1], but
the added values were in the order of 10^8. This completely
rendered ltsgraph useless.

This randomisation may have served a purpose, but as the
purpose was unclear, it has been removed altogether.

It appears to behave much better now.