Attempt to make the graph stable faster.
Up till now the graph would become stable if its center of
gravity was equal to (0,0,0). But typically, it could vibrate
around this value, never becoming exactly equal. This has now
been mended by requiring that the center of gravity should be
at at most a distance of 0.0001 from (0,0,0). Hopefully, this
has the desired effect.