Adapted ltsgraph such that it is more behaved.
Essentially changed the attraction and repulsion function
for graphs in the LTSGraph mode such that it will now strive
to the natural transition length using a force function of
which the derivative is smaller than 1 at the natural transition
length. This should stabilize positioning, but as there are
many more factors influencing positioning, this is unfortunately
not always the case.

For small graphs of less than 200 items (states, labels, etc.)
the use of tree accelleration is switched off.

When unfolding a node in exploration mode, or when changing the
speed of the movement of states, ltsgraph will automatically
set shaping on, when enabled.

Changed a number of labels of buttons and explanation, hopefully
making them clearer.

There is one aspect about which I am still in doubt. When having
state labels on top of each other, a small random force is applied
on them to push them away from each other. This is now a random
value in the range of -0.01 to 0.01. It is unclear how this will
behave.

I hope that ltsgraph is now a little more pleasant to use, but
if not, we should fine tune it further.