Adapted ltsgraph such that it does not use >100% cpu anymore.
Ltsgraph was using 100% cpu cycles on Linux and Apple. This
was due to calling setUpdatesEnabled(true) after each paint event,
which would trigger a next paint event. This meant that ltsgraph
was continuously repainting itself.

This has been replaced by generating an explicit "draw_new_frame"
signal whenever a new layout of the graph would have been calculated.
This draw_new_frame signal triggers an update of the main screen.

It appears that ltsgraph is working decently well, now, but it
may still be that not under all circumstances all internal
events are properly passed around. This then may have to be
mended.

The simulated annealing does not seem to serve a purpose and
will be removed soon.