Fix building with multi-config generators in CMake.
CMake as so-called multi-config generators in which multiple
build types can be built from a single configuration
(in particular Ninja Multi-Config and Visual Studio, Xcode).
For these build types, the configuration is only available at build time.

The trick where ${CMAKE_CFG_INTDIR} is used to determine whether we have a
multi-config build turned out to be error prone (it lead to a define that
kept the variable ${CONFIGURATION} that was not evaluated at build time).

Instead, the generator expression $<CONFIG> is evaluated at build time.