Fix generating data types with parsing 2.0.4
Between parsing 1.6.1 and parsing 2.0.0, there was a change
in the interface of `Token` such that the parser did not need to be passed.
Our code generation was never updated to reflect this change, so this must have
been broken for some time.

In this commit, I remove the parser argument from Token. While debugging,
some other updates were performed, using the logging library, and adding some type hints.