Improve the performance of pbesconstelm
On generated examples (translations of Cordis models into mCRL2),
pbesconstelm took an excessive amount of time. I profiled this using
the VTune profiler, and it turns out most of the time was spent
allocating and deallocating aterms in the edge_condition_traverser.
Most of the allocations and deallocations dominating the running time
can be avoided by using rvalue references instead of lvalue references
to const. This is implemented in this changeset.

On the test case I used to profile this issue, the running time changes
from >10 minutes to 20 seconds.