Only consider simple formulas in modified preSRF rules for or/and
Strengthening conditions not only with simple formulas but also with
predicate variables containing evidence information in the tranformation
to preSRF resulted in right-hand-sides with negated predicate variables.
As many of our tools assume negation does not occur (we seem to assume
positive normal form), this resulted in failures. A notable example is
building a structure graph.

The issue was caught using the random tests in
`random_pbessolvesymbolic-counter-example`.