Normalize PBES obtained from PreSRF.
In most places we expect PBESs to be in positive normal form.
PBES obtained from the PreSRF transformation would generate expressions
of the form !(!Xce && ...), where Xce is a variable with evidence information.
We now normalize this by distributing negation over conjunction and
eliminating double negation.

This issue was detected using the `random_pbessolvesymbolic-counter-example` tests.