Removed dedicated code for assignments in the lineariser.
The code to find free variables in assignments did not take into
account that when variables are missing in assignments, this
may still mean that they are assigned. E.g. in a process
P(x:D)=    ....  P()
the assignment at the rhs means that x is assigned to x, and hence
the variable x should occur free in P(). This has been adapted
in process/add_binding.h.

This allowed to replace the add hoc code find_free_variable_process
in the lineariser, except that process identifiers contain their
parameter_lists, which were inconsistent with the parameter_lists that
the lineariser was using. The lineariser would first rename all
parameters such that each parameter with a certain name would have
a unique type. E.g., x:Bool and x:Nat cannot occur simultaneously
in one process. This renaming is now done at a later moment, not with
an add hoc substitution, but with a general renaming function.

This allowed to remove a substantial amount of code in the lineariser.

Except for the random_pbessolvesymbolic-counter-example test, all
tests succeed. This test is not related to the changed code, and
already failed before this change. A new test has been added.

Some ifdef's regarding skip long tests, which would cause short
tests to be skipped, have been removed.