Resolved problem 1857 where a process was wrongly typechecked.
The following example should not pass typechecking:

act a, A: Nat; a: Bool; init comm({a|a -> A}, a(true)|a(true) . delta);

The behaviour of this process is A(true).delta, but there is no action
A of sort Bool declared. The type checking rules did not check
for this case, and that has now been added.

This resolves #1857.

The example above has been added as an explicit testcase.