Made the typechecking of communication clauses in line with the book.
The book says that if some action ai has some type in a communication
clause a1|...|an -> b then all actions a1...an and b must have that
type. This now looks overkill, and the previous commit of yesterday
did a weaker but sufficient check, but it is better to stay in line
with the book.

Also replaced some auto's by explicit types.