Made splitting conditions in disjunctions more aggressive.
In #1855 it is indicated that conditions in processes such as

sum n:Nat. (n in {1,2 5000000000000000000000}) -> a(n).P(....)

cannot properly be handled by the tools. It would be nice if such
a condition could be eliminated by lpssumelm.

To achieve this, the code to split condition in conjunctions in
lpssumelm has been made a little more aggressive. The condition above
rewrites to:

   (s == 7 && (n==1 || n==2 || n==5000000000000000000000) )

which can now be transformed into a disjunction, splitting summands,
and allowing sum elimination to be applied. The result is:

   (s == 7 && n==1) ||
   (s ==7 && n==2) ||
   (s == 7 && n==5000000000000000000000)

There is now a risk that the resulting disjunction grows exponentially.

Furthermore, some duplicate code has been removed.