Made the new grootejansen25 bisimulation algorithm the default.
The previous default was the jansen/groote/keiren/wijs algorithm
for branching bisimulation. The new one is slightly faster, and
uses slightly less memory.

Besides making this new variant the default, it is also an
explicit option for reduction. The number of branching bisimulation
reduction algorithms in ltsconvert and ltscompare is now a little
too big for comfort. This should be adapted by making some
of the options hidden.