Fix check_strategy.
For winning sets W0 (resp. W1), we only know that the strategy for
player 0 (resp. player 1) is defined. This means that in the solution of
a parity game, there may be vertices, e.g. owned be player 0, for which
no strategy is defined.

When the strategy is applied to the parity game, such vertices become
sinks (and are thus immediately solved), but these new sinks were
not treated as such in check_strategy.

We now recompute the sinks in check_strategy.