Made applying some equivalences a hidden option.
There are many different algorithms that can be selected to reduced
a variant of strong/branching bisimulation in the tools ltsreduce and
ltscompare. As for the user these options are not so interesting, but
clutter the help text, most of these options have now been made
hidden. Using --help-all all arguments that can be used are shown.