Prepared the toolset for an experimental adaptation of the gj-branching bisimulation algorithm.
David Jansen outlined a plan to only calculate the BLC sets when
strictly needed. This could have a substantial effect on both the running
time and the memory requirement of the branching bisimulation algorithm
as the BLC sets are only needed when stabilizing large sets with new
bottom states, but these are rare. Especially, when there are no tau
transitions, the BLC sets are completely unnecessary, making the
algorithm much faster when calculating strong bisimulation.

This new algorithm is called bisim-gj-lazy-BLC (and similar variants for
branching and divergence preserving branching) and are added using
hidden options at the --equivalence flag. A new implementation will be
made in the file liblts_bisim_gj_lazy_BLC.h, but this still needs to be
done.