Repaired a problem in the lineariser.
The lineariser contained a mistake, where it would substitute
variables at some point, without taking the bound variables into
account. This led to havoc on a large specification from Thermo-Fischer.

The problem has now been repaired. This was a consequence of adapting
the substitutions in assignments, which happened after the latest
release.

All tests succeed.