Replaced stack based tau-SCC algorithm by an iterative one.
The stack based algorithm in liblts_scc has been replaced by
an iterative one following the notes in the master thesis of
Wouter Schols, supervised by Wieger Wesselink and Cees Huizing.

This was due to a request by Mieke Massink who had a transition
system that could not be reduced because the tau-SCC algorithm
ran out of the stack on a mac.

Furthermore, (re)enabled determinisation and enabled tau-loop
removal within ltsconvert.

All tests succeed. Regression tests for branching bisimulation
wrt. earlier toolsets, with the old SCC algorithm led to the
same results.