functioning version of lazy-BLC bisimulation minimisation
There are two situations where the bisimulation minimisation algorithm
published in CONCUR 2025 does need the BLC sets. This version contains
alternatives for both situations, that work if the respective block is
a small subblock of an earlier split (because the alternatives do work
proportional to the number of transitions of this block).

- When only the coroutines for AvoidLrg and NewBotSt are unfinished in
  four_way_splitB(), the latter needs to go through the transitions of
  the large splitter to find further NewBotSt-states. Originally, this
  was done by going through the BLC set of the large splitter. The new
  alternative goes through all outgoing transitions of states that are
  not yet in one of the sub-blocks, and if some transition is in the
  large splitter, adds the source state to NewBotSt.

- When a block with new bottom states is stabilised, the original
  algorithm goes through all BLC sets of transitions out of this block
  to decide which refinements are needed. The new alternative goes
  through all outgoing transitions of the block (including those from
  non-bottom states) to find the necessary refinements.

This version still does a lot of heap allocations, and I want to find
a way to reduce that.