Repaired damage done when removing code for assignment substitution.
An example by Flip van Spaendonck showed that removing ad hoc
code in the lineariser to substitute assignments was not done properly.
In order to repair this, the generated function replace_all_variables
was adapted such that it would replace all variables, including those
that are bound and those that occur in the lhs of an assignment.
Furthermore, the typechecker for processes has been adapted to
guarantee that the assignments in a process_identifier_assignment
occur in the same order as the parameters. Tests have been added.

There are some tests that fail, namely the exampletests and
random counter example tests. This seems to be an independent
issue. The exampletest only fails in release mode, and the
problem appears to be related to counterexamples also.