Fix symbolic solving with counterexample information
When partial solving during exploration already solved the PBES,
incomplete vertices (those the were found but not yet explored) were
not properly taken into account for the second solving. This
resulted in inconsistent solutions between the first and second round
of solving.

Note: this commit also removes two checks that were recently added
that are too strict, and cannot be made to hold in general. Also,
it turns out that the root cause of the problem for which they were
added was in the rewriter, and not in reachability and solving of
PBESs.