The pbes rewriters now respect bound variables.
The simplify data rewriter on PBESs did not deal with
bound variables in a correct way. This has now been mended.

It turned out that variables in the right hand sides of mutable
indexed substitutions were not always found correctly, due to
the traverser framework that had a default function apply(const T& x)
that would not do anything if there would not be an explicit
apply function of type T. This default apply function has been removed.
This lead to several changes as in a number of cases, such as finding
action labels in multi-actions that were not alwas correctly found.

A test has been added.