Changed the "marker" for pres's.
The marker in files for pres's was the same as those for pbes's.
This meant that pres tools could operate on pbes files and vice versa.
This has now been mended.

A consequence is that previously generated pres files are not
recognized anymore by the pres tools. Given the current popularity of
pres's this is most likely a minor annoyance.

This resolves bug report #1886.