Consistently compute strategies in different solvers.
Due to the incompatibility between chaining, used in
safe_predecessors_impl, and the computation of strategies in
that same function, we need to ensure that chaining is disabled
when computing strategies.
The class symbolic_parity_game has a flag m_strategy that controls
whether the strategy needs to be computed in safe_predecessors_impl,
and if so, chaining is not used.
When applying partial solvers, this option was not consistently passed
between all solvers, leading to incomplete strategies. This in turn
resulted in incorrect simplified right hand sides being generated in
the second pass of pbessolvesymbolic when generating counterexamples.

The majority of changes in this commit is detailed trace-level debug output
that is useful in debugging the problem.