Made some dependencies of basic data types explicit.
The following declarations -- generated using the code generating
framework -- have been made explicit:

    explicit action_formula(const data::data_expression& x)
    explicit action_formula(const data::untyped_data_parameter& x)
    explicit action_formula(const process::untyped_multi_action& x)
    explicit regular_formula(const action_formulas::action_formula& x)
    explicit regular_formula(const data::data_expression& x)
    explicit state_formula(const data::data_expression& x)
    explicit state_formula(const data::untyped_data_parameter& x)
    explicit pbes_expression(const data::data_expression& x)
    explicit pbes_expression(const data::untyped_data_parameter& x)
    explicit pres_expression(const data::data_expression& x)
    explicit pres_expression(const data::untyped_data_parameter& x)
    explicit process_expression(const data::untyped_data_parameter& x)

The most influential are the following, where types depend on
data_expressions and action formulas:

    explicit action_formula(const data::data_expression& x)
    explicit regular_formula(const action_formulas::action_formula& x)
    explicit regular_formula(const data::data_expression& x)
    explicit state_formula(const data::data_expression& x)
    explicit pbes_expression(const data::data_expression& x)
    explicit pres_expression(const data::data_expression& x)

These now require an explicit atermpp::down_cast<...> to be
transformed to the appropriate type.

The reason for this change is that sometimes these types were implicitly
converted, leading to infinite recursion. For instance when a traverser
would be applied to a pbes_expression, it would figure out the actual
type of the term. If it turned out to be a data expression, it would
use the apply function on data expressions. But if such a function would
not exist, the data_expression would be automatically be converted to
a pbes_expression, after which the traverser framework would find out
that it was a data_expression, leading to an infinite loop.

The adaptation led to having to add atermpp::down_casts at numerous
places. But it also led to the detection of some inapropriate functions
of wrong types being applied. Moreover, some code being put within
code automatically generated by python scripts have been moved
to a place where it would not be overwritten.

All tests succeed.